backend/requirements.txt
# LoopedAI Backend Dependencies

# API Server
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Telegram Bot
python-telegram-bot==21.0

# LLM
requests>=2.31.0

# Config
python-dotenv>=1.0.0

# Supabase (optional — comment out for SQLite-only mode)
supabase>=2.0.0
